+++ resources/js/components/Room/RoomComponent.vue
@@ -1,10 +1,10 @@
 <template>
   <div>
     <b-overlay :show="loading" rounded="sm">
-    <b-card no-body bg-variant="white" class="roomcard" @click="open()">
+    <b-card no-body bg-variant="white" class="room-card" @click="open()">
       <b-card-body class="p-3">
         <b-row>
-          <b-col cols="3" sm="3"><div v-if="type" v-b-tooltip.hover :title="type.description" class="roomicon" :style="{ 'background-color': type.color}">{{type.short}}</div></b-col>
+          <b-col cols="3" sm="3"><div v-if="type" v-b-tooltip.hover :title="type.description" class="room-icon" :style="{ 'background-color': type.color}">{{type.short}}</div></b-col>
           <b-col col><h5 class="mt-2 text-break " style="width: 100%">{{name}}</h5></b-col>
         </b-row>
       </b-card-body>