+++ resources/js/views/settings/SettingsHome.vue
@@ -8,9 +8,9 @@
        <b-col class="p-2">
          <b-card class="h-100">
             <router-link class="stretched-link" :to="{ name: 'settings.application' }">
-            <h2 class="text-center text-success"><i class='fa-solid fa-cogs'></i></h2>
+            <h2 class="text-center text-primary"><i class='fa-solid fa-cogs'></i></h2>
             </router-link>
-            <b-card-title class="text-center text-success">{{ $t('settings.application.title') }}</b-card-title>
+            <b-card-title class="text-center text-primary">{{ $t('settings.application.title') }}</b-card-title>
             <b-card-text>
               {{ $t('settings.application.tileDescription') }}
             </b-card-text>
@@ -21,9 +21,9 @@
        <b-col class="p-2">
          <b-card class="h-100">
             <router-link class="stretched-link" :to="{ name: 'settings.users' }">
-              <h2 class="text-center text-success"><i class='fa-solid fa-users'></i></h2>
+              <h2 class="text-center text-primary"><i class='fa-solid fa-users'></i></h2>
             </router-link>
-            <b-card-title class="text-center text-success">{{ $t('settings.users.title') }}</b-card-title>
+            <b-card-title class="text-center text-primary">{{ $t('settings.users.title') }}</b-card-title>
             <b-card-text>
               {{ $t('settings.users.tileDescription') }}
             </b-card-text>
@@ -34,9 +34,9 @@
        <b-col class="p-2">
           <b-card class="h-100">
             <router-link class="stretched-link" :to="{ name: 'settings.roles' }">
-              <h2 class="text-center text-success"><i class='fa-solid fa-user-tag'></i></h2>
+              <h2 class="text-center text-primary"><i class='fa-solid fa-user-tag'></i></h2>
             </router-link>
-            <b-card-title class="text-center text-success">{{ $t('settings.roles.title') }}</b-card-title>
+            <b-card-title class="text-center text-primary">{{ $t('settings.roles.title') }}</b-card-title>
             <b-card-text>
               {{ $t('settings.roles.tileDescription') }}
             </b-card-text>
@@ -47,9 +47,9 @@
        <b-col class="p-2">
           <b-card class="h-100">
             <router-link class="stretched-link" :to="{ name: 'settings.room_types' }">
-              <h2 class="text-center text-success"><i class='fa-solid fa-tags'></i></h2>
+              <h2 class="text-center text-primary"><i class='fa-solid fa-tags'></i></h2>
             </router-link>
-            <b-card-title class="text-center text-success">{{ $t('settings.roomTypes.title') }}</b-card-title>
+            <b-card-title class="text-center text-primary">{{ $t('settings.roomTypes.title') }}</b-card-title>
             <b-card-text>
               {{ $t('settings.roomTypes.tileDescription') }}
             </b-card-text>
@@ -60,9 +60,9 @@
         <b-col class="p-2">
           <b-card class="h-100">
             <router-link class="stretched-link" :to="{ name: 'settings.servers' }">
-              <h2 class="text-center text-success"><i class='fa-solid fa-server'></i></h2>
+              <h2 class="text-center text-primary"><i class='fa-solid fa-server'></i></h2>
             </router-link>
-            <b-card-title class="text-center text-success">{{ $t('settings.servers.title') }}</b-card-title>
+            <b-card-title class="text-center text-primary">{{ $t('settings.servers.title') }}</b-card-title>
             <b-card-text>
               {{ $t('settings.servers.tileDescription') }}
             </b-card-text>
@@ -73,9 +73,9 @@
         <b-col class="p-2">
           <b-card class="h-100">
             <router-link class="stretched-link" :to="{ name: 'settings.server_pools' }">
-              <h2 class="text-center text-success"><i class='fa-solid fa-th-large'></i></h2>
+              <h2 class="text-center text-primary"><i class='fa-solid fa-th-large'></i></h2>
             </router-link>
-            <b-card-title class="text-center text-success">{{ $t('settings.serverPools.title') }}</b-card-title>
+            <b-card-title class="text-center text-primary">{{ $t('settings.serverPools.title') }}</b-card-title>
             <b-card-text>
               {{ $t('settings.serverPools.tileDescription') }}
             </b-card-text>